Allow user to implement their own error logging
