+++ example/server.js
@@ -19,6 +19,15 @@ fs.readdirSync(path.join(__dirname, "/resources")).filter(function(filename) {
   return path.join(__dirname, "/resources/", filename);
 }).forEach(require);
 
+jsonApi.onError(function(request, error) {
+  var errorDetails = error.stack.split("\n");
+  console.error(JSON.stringify({
+    request: request,
+    error: errorDetails.shift(),
+    stack: errorDetails
+  }));
+});
+
 jsonApi.start();
 server.start = jsonApi.start;
 server.close = jsonApi.close;
