+++ lib/jsonApi.js
@@ -79,6 +79,10 @@ jsonApi.define = function(resourceConfig) {
   })));
 };
 
+jsonApi.onError = function(errHandler) {
+  jsonApi._errHandler = errHandler;
+};
+
 jsonApi.start = function() {
   routes.register();
   router.listen(jsonApi._apiConfig.port);
