+++ lib/routes/zerrorHandler.js
@@ -1,6 +1,7 @@
 "use strict";
 var errorHandler = module.exports = { };
 
+var jsonApi = require("../jsonApi.js");
 var helper = require("./helper.js");
 var router = require("../router.js");
 
@@ -8,12 +9,9 @@ var router = require("../router.js");
 errorHandler.register = function() {
   router.bindErrorHandler(function(request, res, error) {
 
-    var errorDetails = error.stack.split("\n");
-    console.error(JSON.stringify({
-      request: request,
-      error: errorDetails.shift(),
-      stack: errorDetails
-    }));
+    if (jsonApi._errHandler) {
+      jsonApi._errHandler(request, error);
+    }
 
     return helper.handleError(request, res, {
       status: "500",
